+++ ext/nokogiri/xml_cdata.c
@@ -5,6 +5,9 @@
  *  new(document, content)
  *
  * Create a new CDATA element on the +document+ with +content+
+ *
+ * If +content+ cannot be implicitly converted to a string, this method will
+ * raise a TypeError exception.
  */
 static VALUE new(int argc, VALUE *argv, VALUE klass)
 {
@@ -14,23 +17,24 @@ static VALUE new(int argc, VALUE *argv, VALUE klass)
   VALUE content;
   VALUE rest;
   VALUE rb_node;
+  const xmlChar *content_str;
+  int content_str_len;
 
   rb_scan_args(argc, argv, "2*", &doc, &content, &rest);
 
   Data_Get_Struct(doc, xmlDoc, xml_doc);
 
-  node = xmlNewCDataBlock(
-      xml_doc->doc,
-      NIL_P(content) ? NULL : (const xmlChar *)StringValuePtr(content),
-      NIL_P(content) ? 0 : (int)RSTRING_LEN(content)
-  );
+  content_str = NIL_P(content) ? NULL : (const xmlChar *)StringValueCStr(content);
+  content_str_len = (content_str == NULL) ? 0 : strlen(content_str);
+
+  node = xmlNewCDataBlock(xml_doc->doc, content_str, content_str_len);
 
   nokogiri_root_node(node);
 
   rb_node = Nokogiri_wrap_xml_node(klass, node);
   rb_obj_call_init(rb_node, argc, argv);
 
-  if(rb_block_given_p()) rb_yield(rb_node);
+  if(rb_block_given_p()) { rb_yield(rb_node); }
 
   return rb_node;
 }